requirements.txt
python-chess>=1.9.0
requests>=2.31.0
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
cairosvg>=2.7.1
Pillow>=10.0.0
beautifulsoup4>=4.12.0 